Move binary encoding/decoding to syntax/
