+++ dhall/src/semantics/phase/parse.rs
@@ -3,9 +3,9 @@ use std::io::Read;
 use std::path::Path;
 
 use crate::semantics::error::Error;
-use crate::semantics::phase::binary;
 use crate::semantics::phase::resolve::ImportRoot;
 use crate::semantics::phase::Parsed;
+use crate::syntax::binary;
 use crate::syntax::parse_expr;
 
 pub(crate) fn parse_file(f: &Path) -> Result<Parsed, Error> {
